+++ docs/guide/counter/cmd/countercli/main.go
@@ -14,12 +14,12 @@ import (
 	"github.com/tendermint/tmlibs/cli"
 
 	bcmd "github.com/tendermint/basecoin/cmd/basecli/commands"
-	coincmd "github.com/tendermint/basecoin/cmd/basecoin/commands"
+	bcount "github.com/tendermint/basecoin/docs/guide/counter/cmd/countercli/commands"
 )
 
 // BaseCli represents the base command when called without any subcommands
 var BaseCli = &cobra.Command{
-	Use:   "basecli",
+	Use:   "countercli",
 	Short: "Light client for tendermint",
 	Long: `Basecli is an version of tmcli including custom logic to
 present a nice (not raw hex) interface to the basecoin blockchain structure.
@@ -33,16 +33,25 @@ func main() {
 	commands.AddBasicFlags(BaseCli)
 
 	// Prepare queries
-	pr := proofs.RootCmd
-	// These are default parsers, but optional in your app (you can remove key)
-	pr.AddCommand(proofs.TxCmd)
-	pr.AddCommand(proofs.KeyCmd)
-	pr.AddCommand(bcmd.AccountQueryCmd)
+	proofs.RootCmd.AddCommand(
+		// These are default parsers, optional in your app
+		proofs.TxCmd,
+		proofs.KeyCmd,
+		bcmd.AccountQueryCmd,
 
-	// you will always want this for the base send command
+		// XXX IMPORTANT: here is how you add custom query commands in your app
+		bcount.CounterQueryCmd,
+	)
+
+	// Prepare transactions
 	proofs.TxPresenters.Register("base", bcmd.BaseTxPresenter{})
-	tr := txs.RootCmd
-	tr.AddCommand(bcmd.SendTxCmd)
+	txs.RootCmd.AddCommand(
+		// This is the default transaction, optional in your app
+		bcmd.SendTxCmd,
+
+		// XXX IMPORTANT: here is how you add custom tx construction for your app
+		bcount.CounterTxCmd,
+	)
 
 	// Set up the various commands to use
 	BaseCli.AddCommand(
@@ -50,12 +59,11 @@ func main() {
 		commands.ResetCmd,
 		keycmd.RootCmd,
 		seeds.RootCmd,
-		pr,
-		tr,
+		proofs.RootCmd,
+		txs.RootCmd,
 		proxy.RootCmd,
-		coincmd.VersionCmd,
 	)
 
-	cmd := cli.PrepareMainCmd(BaseCli, "BC", os.ExpandEnv("$HOME/.basecli"))
+	cmd := cli.PrepareMainCmd(BaseCli, "CTL", os.ExpandEnv("$HOME/.countercli"))
 	cmd.Execute()
 }