+++ scripts/dist_build.sh
@@ -18,11 +18,6 @@ GIT_IMPORT="github.com/tendermint/basecoin/version"
 XC_ARCH=${XC_ARCH:-"386 amd64 arm"}
 XC_OS=${XC_OS:-"solaris darwin freebsd linux windows"}
 
-# Delete the old dir
-echo "==> Removing old directory..."
-rm -rf build/pkg
-mkdir -p build/pkg
-
 # Make sure build tools are available.
 make tools
 
@@ -30,7 +25,7 @@ make tools
 make get_vendor_deps
 
 # Build!
-echo "==> Building..."
+echo "==> Building basecoin..."
 "$(which gox)" \
 		-os="${XC_OS}" \
 		-arch="${XC_ARCH}" \
@@ -39,6 +34,16 @@ echo "==> Building..."
 		-output "build/pkg/{{.OS}}_{{.Arch}}/basecoin" \
 		-tags="${BUILD_TAGS}" \
 		github.com/tendermint/basecoin/cmd/basecoin
+
+echo "==> Building basecli..."
+"$(which gox)" \
+    -os="${XC_OS}" \
+    -arch="${XC_ARCH}" \
+    -osarch="!darwin/arm !solaris/amd64 !freebsd/amd64" \
+    -ldflags "-X ${GIT_IMPORT}.GitCommit='${GIT_COMMIT}' -X ${GIT_IMPORT}.GitDescribe='${GIT_DESCRIBE}'" \
+    -output "build/pkg/{{.OS}}_{{.Arch}}/basecli" \
+    -tags="${BUILD_TAGS}" \
+    github.com/tendermint/basecoin/cmd/basecli
 
 # Zip all the files.
 echo "==> Packaging..."